requirements.txt
python-dotenv
pytesseract
pillow
pandas
pyyaml
openai
opencv-python-headless
pytest
regex
